Changing the translation and utils file.
